+++ docs/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 * :bug:`-` Users will be able to see the address of each account within an xpub.
 * :bug:`-` An exception in the last decoding step will no longer stop transaction decoding in rotki.
 * :bug:`-` Failed paraswap v6 swaps will no longer fail to decode in rotki.
+* :bug:`-` Users should now be again able to edit the underlying tokens for any pool tokens.
 
 * :release:`1.38.1 <2025-03-14>`
 * :feature:`9385` Users will see the default label using the validator index when exporting validators.